+++ src/test/java/es/udc/fi/dc/irlab/ppc/hcomputation/TestH3Mapper.java
@@ -1,3 +1,19 @@
+/**
+ * Copyright 2013 Daniel Valcarce Silva
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package es.udc.fi.dc.irlab.ppc.hcomputation;
 
 import static org.junit.Assert.assertEquals;
@@ -5,7 +21,7 @@ import static org.junit.Assert.assertEquals;
 import java.io.IOException;
 import java.util.List;
 
-import org.apache.hadoop.io.LongWritable;
+import org.apache.hadoop.io.IntWritable;
 import org.apache.hadoop.io.NullWritable;
 import org.apache.hadoop.mrunit.mapreduce.MapDriver;
 import org.apache.hadoop.mrunit.types.Pair;
@@ -16,18 +32,20 @@ import org.apache.mahout.math.VectorWritable;
 import org.junit.Before;
 import org.junit.Test;
 
+import es.udc.fi.dc.irlab.ppc.hcomputation.H3Mapper;
+
 public class TestH3Mapper {
 
-    private MapDriver<LongWritable, VectorWritable, NullWritable, MatrixWritable> mapDriver;
+    private MapDriver<IntWritable, VectorWritable, NullWritable, MatrixWritable> mapDriver;
 
     @Before
     public void setup() {
-	mapDriver = new MapDriver<LongWritable, VectorWritable, NullWritable, MatrixWritable>();
+	mapDriver = new MapDriver<IntWritable, VectorWritable, NullWritable, MatrixWritable>();
     }
 
     @Test
     public void testMap() throws IOException {
-	LongWritable inputKey = new LongWritable(1);
+	IntWritable inputKey = new IntWritable(1);
 	Vector inputVector = new DenseVector(new double[] { 1.0, 2.0, 3.0 });
 	VectorWritable inputValue = new VectorWritable(inputVector);
